feat(api): create separate api provider feat(ssr): main and detail ssr data call
