+++ src/app/anime/[id]/watch/[epsId]/page.tsx
@@ -14,7 +14,7 @@ import Loading from "./loading";
 import { Badge } from "@/components/ui/badge";
 import { IAnimeResult } from "@consumet/extensions";
 import Player from "@/components/player";
-import BackHomeButton from "@/components/back-home-button";
+import PreviousButton from "@/components/back-home-button";
 
 export default function WatchPage() {
    const { id: animeId, epsId: episodeId } = useParams();
@@ -36,7 +36,7 @@ export default function WatchPage() {
    return (
       <div className="min-h-screen bg-black text-white flex flex-col">
          {/* Video Player Container */}
-         <BackHomeButton />
+         <PreviousButton />
          <Player
             epsId={episodeId?.toString()}
             animeId={animeId?.toString()}
