+++ src/components/hooks/useAnimes.ts
@@ -1,19 +1,18 @@
-"use client";
-import { useQuery } from "@apollo/client";
+// import { useQuery } from "@apollo/client";
 import { ANIMES_QUERY } from "./queries/main.query";
 import animeStore from "@/lib/stores/animes.store";
 import { ProviderEnum } from "@/lib/anime-provider/provider.interfaces";
-import { useEffect } from "react";
+// import { useEffect } from "react";
+import { createClient } from "@/lib/graphql/client";
 
-export default function useMainAnimes() {
-   const animes = animeStore().main;
-   const page = animeStore().page;
-   const recent_page = animeStore().recent_page;
-   const top_page = animeStore().top_page;
-   const movie_page = animeStore().movie_page;
-   const popular_page = animeStore().popular_page;
-   const provider = animeStore().provider;
-   const setMain = animeStore().setMain;
+export default async function useMainAnimes() {
+   const page = animeStore.getState().page;
+   const recent_page = animeStore.getState().recent_page;
+   const top_page = animeStore.getState().top_page;
+   const movie_page = animeStore.getState().movie_page;
+   const popular_page = animeStore.getState().popular_page;
+   const provider = animeStore.getState().provider;
+   const setMain = animeStore.getState().setMain;
 
    const pagination =
       provider === ProviderEnum.ANIDRV
@@ -25,14 +24,23 @@ export default function useMainAnimes() {
               popular_page,
            };
 
-   const { loading, data, error } = useQuery(ANIMES_QUERY, {
-      variables: { ...pagination },
-      pollInterval: 10_000,
+   // const { loading, data, error, networkStatus } = useQuery(ANIMES_QUERY, {
+   //    variables: { ...pagination },
+   // });
+
+   // useEffect(() => {
+   //    if (data?.main) setMain(data.main);
+   // }, [data]);
+   const client = createClient();
+   const { loading, data, error } = await client.query({
+      query: ANIMES_QUERY,
+      variables: pagination,
    });
 
-   useEffect(() => {
-      if (data?.main) setMain(data.main);
-   }, [data]);
+   if (data.main) setMain(data.main);
+   const animes = animeStore.getState().main;
+
+   console.log(animes);
 
    return { loading, animes, error };
 }
