+++ src/components/hooks/useDetail.ts
@@ -1,21 +1,18 @@
-"use client";
-import { useQuery } from "@apollo/client";
 import animeStore from "@/lib/stores/animes.store";
-import { useEffect } from "react";
 import { DETAIL_QUERY } from "./queries/detail.query";
+import { createClient } from "@/lib/graphql/client";
 
-export default function useDetailAnime(id: string) {
-   const detail = animeStore().detail;
-   const setMain = animeStore().setDetail;
+export default async function useDetailAnime(id: string) {
+   const setDetail = animeStore.getState().setDetail;
+   const client = createClient();
 
-   const { loading, data, error } = useQuery(DETAIL_QUERY, {
+   const { loading, data, error } = await client.query({
+      query: DETAIL_QUERY,
       variables: { id },
-      pollInterval: 10_000,
    });
 
-   useEffect(() => {
-      if (data?.detail) setMain(data.detail);
-   }, [data]);
+   if (data?.detail) setDetail(data.detail);
+   const detail = animeStore.getState().detail;
 
    return { loading, detail, error };
 }
