+++ src/lib/api/api.interface.ts
@@ -1,0 +1,7 @@
+export interface IStreamData {
+   animeId: string;
+   episodeId: string;
+   segment: string;
+   data: string; // Buffer to base64
+   headers: Record<string, string>;
+}
