+++ src/lib/stores/auth.store.ts
@@ -1,4 +1,3 @@
-"use client";
 import { create } from "zustand";
 import { createJSONStorage, persist } from "zustand/middleware";
 import { IAuthData, IAuthStore } from "./interfaces/auth.interfaces";
